+++ BALSAMIC/utils/constants.py
@@ -362,6 +362,22 @@ REFERENCE_FILES = {
             "output_file": "delly_exclusion.tsv",
             "output_path": "genome",
         },
+        "ascat_gccorrection": {
+            "url": "https://raw.githubusercontent.com/Clinical-Genomics/reference-files/35465e2644f76f2d59427a9b379d34ecea71f259/cancer/references/hg38_SnpGcCorrections.tsv.gz",
+            "file_type": "text",
+            "gzip": True,
+            "genome_version": "hg38",
+            "output_file": "hg38_SnpGcCorrections.tsv",
+            "output_path": "genome",
+        },
+        "ascat_chryloci": {
+            "url": "https://raw.githubusercontent.com/Clinical-Genomics/reference-files/35465e2644f76f2d59427a9b379d34ecea71f259/cancer/references/hg38_Y.loci",
+            "file_type": "text",
+            "gzip": False,
+            "genome_version": "hg38",
+            "output_file": "hg38_Y.loci",
+            "output_path": "genome",
+        },
     },
     "hg19": {
         "reference_genome": {