+++ CHANGELOG.rst
@@ -18,6 +18,10 @@ Fixed:
 
 * Fixed context for Dockerfile for release content #720
 
+Added:
+
+* ``ascatngs`` hg38 references to constants #683 
+
 [8.0.0]
 -------